third_party/WebKit/LayoutTests/http/tests/flags/exclusive.html

<!DOCTYPE html>
<title>Origin Flags API: Exclusive Mode</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async t => {
  const granted = [];
  function grant(n) { return () => { granted.push(n); }; }

  await Promise.all([
    requestFlag(['a', 'b'], 'exclusive').then(grant(1)),
    requestFlag(['b', 'c'], 'exclusive').then(grant(2)),
    requestFlag(['c'], 'exclusive').then(grant(3))
  ]);
  assert_array_equals(granted, [1, 2, 3]);
}, 'Flag requests are granted in order');

promise_test(async t => {
  let unblock;
  const blocked = new Promise(r => { unblock = r; });

  const granted = [];
  function grant(n) { return () => { granted.push(n); }; }

  const f = await requestFlag('a', 'exclusive');
  f.waitUntil(blocked);

  const ps = Promise.all([
    // This will be blocked.
    requestFlag(['a', 'b'], 'exclusive').then(grant(1)),
    // But this should be grantable immediately.
    requestFlag(['c', 'd'], 'exclusive').then(grant(2))
  ]);

  unblock();

  await ps;
  assert_array_equals(granted, [2, 1]);
}, 'Requests with non-overlapping scopes can be granted');

</script>
